// src/app/model/coleccion.model.ts
export interface ModelColecciones {
    id?: string
    nombre?: string
    descripcion?: string
    descripcionL?: string
    urlImg?: string
    idArea?: number
    nombreArea?: string
    idBitacora?: number
    fechaInsercion?: Date
    idUsuarioInsercion?: number
    fechaUltMod?: Date
    idUsuarioUltMod?: number
}